Almost fully working
