+++ client/coral-framework/helpers/strings.js
