--- client/coral-reaction/components/style.css
@@ -1,30 +1,0 @@
-.like {
-    display: inline-block;
-}
-
-.button {
-    color: #2a2a2a;
-    margin: 5px 10px 5px 0px;
-    background: none;
-    padding: 0px;
-    border: none;
-    font-size: inherit;
-
-    &:hover {
-         color: #767676;
-         cursor: pointer;
-     }
-
-    &.liked {
-         color: rgb(0,134,227);
-
-        &:hover {
-         color: rgb(0,134,227);
-         cursor: pointer;
-        }
-    }
-}
-
-.icon {
-    padding: 0 5px;
-}
