--- client/coral-reaction/index.js
@@ -1,3 +1,0 @@
-import CoralReaction from './containers/CoralReaction';
-
-export default CoralReaction;
